wip: split up divider and reroll component
